Persist parse and signature work across runs, and fix two row-4 accuracy defects
Analysis now reuses parsed trees, fingerprints, and MinHash signatures from
an on-disk store, so a second run over an unchanged tree does no tree-sitter
work and an edit re-analyses only the edited file. Warm runs are byte-equal
to cold runs apart from cache_stats.

Blob trust is a prerequisite to decode: every blob carries a BLAKE3 digest
binding it to the full address that wrote it, verified before any payload
byte is read. The read is bounded on the read itself, allocations are
fallible, and a global decoded-node budget bounds a wide-but-shallow tree
the depth guard cannot see.

Retention deletes nothing under a 2 GiB budget: orphans are the revert-reuse
set, and another tool version's partition may belong to a second binary
sharing the workspace.

Two accuracy defects surfaced while pinning the LSH-only route. Taxonomy row
4 was implemented for C# only, so every other language reported a fully
duplicated pair as zero duplication. Routing it everywhere then let framework
scaffolding reach the act-now bucket, so row 4 now demotes cross-file spreads
and clusters the content pass could not measure. Content-gating it instead
was measured and rejected: it demotes a renamed Type-3 clone, which neither
of that gate's populations can evaluate without positional alignment.
