+++ crates/deslop-core/src/report_render.rs
@@ -18,7 +18,7 @@ use crate::{
     buckets::{
         bucket_labels, classify_signals, content_gated_signals, has_saturating_shape_evidence,
         is_lsh_only_nearmiss, is_structural_only_signals, lacks_content_support, ClusterKind,
-        CONTENT_PROMOTE_FLOOR, CONTENT_SUPPORT_FLOOR, LITERAL_TABLE_MIN_FRACTION,
+        CONTENT_PROMOTE_FLOOR, LITERAL_TABLE_MIN_FRACTION,
     },
     cluster::Cluster,
     cluster_filters::ParseCache,
@@ -336,6 +336,10 @@ pub(crate) fn report_bucket_kind(
 /// with no semantic backing routes to the demoted structural-only tier
 /// — [`ClusterKind::LooselySimilar`] for the cross-file scaffolding
 /// spread, [`ClusterKind::StructuralOnly`] otherwise.
+///
+/// The anchor-free near-miss ([CLONE-BUCKETS-ROUTING] row 4) is decided
+/// *before* that, by [`route_anchor_free`], and on different evidence —
+/// it has no shape match for this gate's populations to measure against.
 fn route_shape_identical(
     kind: ClusterKind,
     signals: ReportSignals,
@@ -345,24 +349,14 @@ fn route_shape_identical(
     if !matches!(
         kind,
         ClusterKind::NearlyIdentical | ClusterKind::StructuralOnly
-    ) || !has_saturating_shape_evidence(signals)
-    {
+    ) {
         return kind;
     }
-    // [CLONE-BUCKETS-ROUTING] row 4 needs *content* proof, because it has
-    // nothing else: `structural ≤ 0.01` means no shape matched at all, so
-    // a normalised-token estimate is the cluster's only evidence. Weak or
-    // absent agreement leaves a hint, not a verdict —
-    // [`ClusterKind::LooselySimilar`], which the renderer hides — and
-    // never [`ClusterKind::StructuralOnly`], which would claim a shape
-    // match that does not exist. `proven_support` (not `support`) is
-    // deliberate: the anchored routes may take an unmeasured cluster on
-    // trust because their Merkle equality is itself proof, while here an
-    // unmeasured cluster is simply unproven. Without this rule the #108
-    // JSON-schema pair (`structural=0.00, token_jaccard=0.96`, no content
-    // measurement) is routed act-now on no evidence whatsoever.
-    if is_lsh_only_nearmiss(signals) && content.proven_support() < CONTENT_SUPPORT_FLOOR {
-        return ClusterKind::LooselySimilar;
+    if let Some(demoted) = route_anchor_free(signals, content, members) {
+        return demoted;
+    }
+    if !has_saturating_shape_evidence(signals) {
+        return kind;
     }
     let shape_only = is_structural_only_signals(signals) || lacks_content_support(signals, content);
     if !shape_only {
@@ -389,6 +383,51 @@ fn route_shape_identical(
         return ClusterKind::LooselySimilar;
     }
     ClusterKind::StructuralOnly
+}
+
+/// Demotion for [CLONE-BUCKETS-ROUTING] **row 4** — the anchor-free
+/// near-miss — or `None` to leave the routing alone. `structural ≤ 0.01`
+/// means no shape matched at all, so a normalised-token estimate is the
+/// cluster's only evidence, and two shapes of cluster carry that estimate
+/// without earning an act-now verdict:
+///
+/// - **A cross-file spread** (3+ members over 3+ files) is the #134
+///   scaffolding pattern arriving through the token door instead of the
+///   structural one. Six distinct Flutter widgets read `structural=0.00,
+///   token_jaccard=0.93` over whole-file spans whose `build` bodies share
+///   nothing, because the framework-mandated declaration is most of each
+///   file (#331). **A genuine clone family of that width is demoted to a
+///   hint too** — the same trade [`is_cross_file_scaffolding`] already
+///   makes for shape-identical spreads, for the same reason, and it is a
+///   trade rather than a free win. Two narrower discriminators were
+///   measured and rejected: the content gate (see
+///   [`has_saturating_shape_evidence`]) and
+///   [`ContentEvidence::substance_varies`], both of which demote
+///   `csharp-type3` — a genuine renamed Type-3 pair — because neither can
+///   evaluate a rename across misaligned shapes. Narrowing this rule
+///   needs a discriminator that survives that fixture.
+/// - **An unmeasured cluster**, where the content pass could not compare
+///   two members at all. The anchored routes may take one on trust
+///   because their Merkle equality is itself proof; row 4 has no such
+///   signal, so unmeasured there means *nothing is known*. The #108
+///   JSON-schema pair (`structural=0.00, token_jaccard=0.96`) would
+///   otherwise be routed act-now on no evidence whatsoever.
+///
+/// The destination is [`ClusterKind::LooselySimilar`] — a hint the
+/// renderer hides — and never [`ClusterKind::StructuralOnly`], which
+/// would claim a shape match `structural = 0.00` says does not exist.
+///
+/// A *pair* that the content pass did measure is left alone even when its
+/// agreement is low: that is the renamed Type-3 clone
+/// ([`has_saturating_shape_evidence`] documents the 0.19 measurement),
+/// which this gate's populations are structurally unable to vouch for.
+fn route_anchor_free(
+    signals: ReportSignals,
+    content: ContentEvidence,
+    members: &[Fingerprint],
+) -> Option<ClusterKind> {
+    let unearned = !content.measured || is_cross_file_scaffolding(members);
+    (is_lsh_only_nearmiss(signals) && unearned).then_some(ClusterKind::LooselySimilar)
 }
 
 /// Returns true when a structural-only cluster spans enough distinct
