+++ docs/incremental-persistence-regression-audit.md
@@ -30,12 +30,20 @@ The LSH-only fixture finding 6 asked for did its job twice: it exposed a false n
 
 **False negative — row 4 was implemented for one language (gh #390).** An authored Python pair (`structural = 0.00`, `token_jaccard = 0.9297`, endpoints past every LSH-only survival floor) rendered **zero** duplication: `classify_signals` had no row-4 arm, so the triple fell to `LooselySimilar`, which the renderer hides, while `report_render::is_csharp_lsh_type3_near_miss` patched the row for C# members only. The carve-out is dissolved into the router — row 4 now routes in every language, as the spec always said.
 
-**False positive — row 4 reached an act-now bucket on no evidence.** Row 4 admits on token overlap *alone*, so it must pass through [FUSION-CONTENT-GATE] like every other shape-saturating route; it did not, because the gate's trigger only knew the higher `SATURATING_TOKEN_FLOOR` (0.95) and row 4 admits at 0.90. Six distinct Flutter widgets measure `structural = 0.00, token_jaccard = 0.93` over whole-file spans whose `build` bodies share nothing — the framework-mandated declaration is most of each file — and were reported at `fused = 0.93` as "nearly identical, review the locations" (#331). Two changes close it:
+**False positive — row 4 reached an act-now bucket on no evidence.** Six distinct Flutter widgets measure `structural = 0.00, token_jaccard = 0.93` over whole-file spans whose `build` bodies share nothing — the framework-mandated declaration is most of each file — and once row 4 routed in every language they were reported at `fused = 0.93` as "nearly identical, review the locations" (#331). The same door admitted #108's JSON-schema pair at `token_jaccard = 0.96`.
 
-1. `has_saturating_shape_evidence` now fires on the anchor-free route at its own floor, so row 4 faces the content gate.
-2. An anchor-free cluster whose *proven* content support is below the floor routes to `LooselySimilar` — never `StructuralOnly`, which would claim a shape match that `structural = 0.00` says does not exist. "Proven" is the load-bearing word: the gate's "never demote on a missing measurement" default is safe only where another signal proves duplication, and an exact Merkle match does. Row 4 has no such signal, so an *unmeasured* cluster there is unproven, not entitled to the benefit of the doubt. Without this, the #108 JSON-schema pair (`structural = 0.00, token_jaccard = 0.96`, no content measurement) is routed act-now on no evidence whatsoever.
+The obvious fix — send row 4 through [FUSION-CONTENT-GATE] like every other route that rests on the normalised representation — is **wrong, and was measured to be wrong before being discarded.** Both of that gate's populations assume the members align position for position, and `structural ≤ 0.01` says the shapes differ. Against the `csharp-type3` fixture — a genuine Type-3 clone with *every* identifier renamed and one extra statement — agreement collapses to 0.19 (the literals) and rename consistency to 0.00, because the extra statement destroys the alignment the rename proof needs. Gating row 4 on content demoted that pair to `structural_only` at `fused = 0.17`: a false negative on the most valuable clone class there is, traded for #331's precision. `cli/detection.rs::detects_type3_clone_in_csharp_fixture` caught it.
 
-The two directions are pinned against each other: `lsh_only_nearmiss_recall.rs` asserts the genuine reorder clone keeps `fused ≥ 0.85` — so a precision fix that widens the gate over real duplicates fails — and `issue_331_336_shape_only_saturation.rs` asserts the scaffold family does not reach that line.
+`ContentEvidence::substance_varies` fails the same fixture for the same reason — it reads `true` for a consistent rename whose extra statement breaks the alignment — so it is not a narrower substitute either. Any future narrowing needs a discriminator that survives `csharp-type3`.
+
+Row 4 is routed on cluster **spread** instead, and only two shapes are demoted:
+
+1. **A cross-file spread** (3+ members over 3+ files) — the #134 scaffolding pattern arriving through the token door instead of the structural one. This is the widget family. **This is a trade, not a free win:** a genuine clone family that wide is demoted to a hint too, exactly as `is_cross_file_scaffolding` already does for shape-identical spreads, for the same stated reason. It is recorded in the function doc, the taxonomy row, and the PR body rather than left implicit.
+2. **An unmeasured cluster**, where the content pass could not compare two members at all. The anchored routes may take one on trust because their Merkle equality is itself proof; row 4 has no such signal, so unmeasured there means *nothing is known*. This is #108. `ContentEvidence` gained a `measured` flag to express it, because "measured full agreement" and "nothing was measured" were previously the same value.
+
+Both demote to `LooselySimilar`, which the renderer hides — never `StructuralOnly`, which would claim a shape match `structural = 0.00` says does not exist. A measured *pair* is left alone even at low agreement: that is the renamed Type-3 clone.
+
+Four suites now pin the four corners, and no two can be satisfied by loosening a threshold: `lsh_only_nearmiss_recall.rs` (genuine LSH-only pair keeps `fused ≥ 0.85`), `cli/detection.rs` (the renamed C# pair still surfaces at `structural = 0.00`), `issue_331_336_shape_only_saturation.rs` (the scaffold family stays below the act-now line), and `issue_98_99_108_120_122_thresholds.rs` (unmeasured noise stays out of the ranked report entirely).
 
 ## Retention, and what a bound means here
 
@@ -51,7 +59,7 @@ Two ceilings are deliberate ceilings, not eliminated risks, and both degrade to 
 ## Verification
 
 - `make ci` — full workspace gate (fmt, clippy `-D warnings`, tests with coverage thresholds, build).
-- Accuracy suites re-run after every fix in this round: `lsh_only_nearmiss_recall` 2/2 · `issue_331_336_shape_only_saturation` 3/3 · `issue_98_99_108_120_122_thresholds` 1/1 · `cache_retention` 3/3 · `fpcache` unit group 19/19 (13 blob/cache + 6 retention).
+- Accuracy suites re-run after every fix in this round: `lsh_only_nearmiss_recall` 2/2 · `issue_331_336_shape_only_saturation` 3/3 · `issue_98_99_108_120_122_thresholds` 1/1 · `cache_retention` 3/3 · `fpcache` unit group 21/21 (14 in `fpcache::tests`, 7 in `fpcache::retention::tests`).
 - Incremental suites: `signature_reuse` 4/4 · `cache_blob_integrity` 4/4 · `incremental_equivalence` 6/6 · `incremental_multilang_golden` 3/3 · `incremental_multilang_matrix` 4/4 · `live_session_status` 3/3.
 - Real-repo lifecycle proof on `deslop-core/src` (135 files, release): cold 1.03 s (0 hit / 135 miss, 29,120 signatures built) → warm 0.43 s (135 hit / 0 miss, **0 built / 29,120 reused**) → edit one file 0.43 s (134 hit / 1 miss, only 165 signatures rebuilt) → revert full-hits. Cold, warm, and revert are byte-equal to the `--no-incremental` report modulo `cache_stats`; the store survives process death (136 blobs / 34 MiB).
 - Pinned tokio benchmark (release, default `min-nodes`, `--embeddings off`): `--no-incremental` 6.45 s / 1,532 MB · cold 6.48 s / 1,550 MB · **warm 3.31 s / 1,495 MB** · one-file edit 3.39 s · revert 3.42 s · store 185.8 MiB / 759 blobs.
