+++ docs/specs/fusion.md
@@ -116,16 +116,18 @@ routing line) saturates on shape
 matches exactly as `structural` does — the surviving flutter/flutter #331
 cluster read `structural=0.62, token_jaccard=0.98, fused=1.00` because
 transitive closure mixed structural and LSH pairs. The gate therefore fires on
-*either* saturating signal — and on the **anchor-free** near-identical route
-([taxonomy.md §CLONE-BUCKETS-ROUTING](taxonomy.md#clone-buckets-routing) row 4)
-at its own lower floor, `admission.lsh_only_min_jaccard`, because that row
-reaches an act-now bucket on token overlap **alone**. Six distinct Flutter
-widget files measure `structural=0.00, token_jaccard=0.93` over whole-file
-spans whose `build` bodies share nothing (#331): the framework-mandated
-declaration is most of each file, so the echo clears the row-4 floor while
-sitting in the band the `saturating_token_floor` test cannot see. Any route
-into `nearly_identical` that rests on the normalised representation alone owes
-the gate its content measurement. Shape-mismatched members have no positional
+*either* saturating signal.
+
+**The gate stops at the anchor-free route.** Row 4 of
+[taxonomy.md §CLONE-BUCKETS-ROUTING](taxonomy.md#clone-buckets-routing) is
+deliberately outside it. Both populations below assume the members align
+position for position, and `structural ≤ 0.01` says the shapes differ — so
+against a genuine Type-3 clone whose identifiers are all renamed and whose
+bodies differ by one statement (`csharp-type3`), agreement collapses to the
+literals (0.19) and rename consistency to 0.00, because the extra statement
+destroys the alignment the rename proof needs. Gating row 4 here would demote
+the renamed near-miss, the most valuable clone class there is. Row 4 is routed
+on cluster *spread* instead — see the taxonomy row. Shape-mismatched members have no positional
 alignment, so their agreement is the key-set Jaccard of their content keys — a
 genuine Type-3 near-miss shares nearly all of them; renamed scaffolding shares
 few. The verbatim guard is proportional
