+++ docs/specs/taxonomy.md
@@ -81,7 +81,9 @@ The canonical signal thresholds that map a cluster's `(structural, token_jaccard
 1. **Evidence-free** (`is_structural_only_signals`) — token and embedding support both below `deslop-core::buckets::STRUCTURAL_ONLY_MAX_SUPPORT` (0.05, tolerating MinHash collision noise). Shape is the only positive signal; the #197 REST settings family is the canonical case.
 2. **Content-gated** (`lacks_content_support`, [FUSION-CONTENT-GATE]) — the deterministic signals saturate *by construction* and therefore prove nothing about content. The token LSH pass hashes the same normalised representation the structural pass does, so an anchor-poor Type-2 rename reads `structural = 1.00, token_jaccard = 1.00` while its raw collapsed leaves disagree and no literal-anchored substitution explains them. `ContentEvidence::support()` below `CONTENT_SUPPORT_FLOOR` routes it here.
 
-**Row 4 is content-gated too.** It reaches an act-now bucket on token overlap *alone*, and that overlap is measured over the normalised stream, so the same argument applies at row 4's own lower floor: six distinct Flutter widgets whose `build` bodies share nothing measure `structural = 0.00, token_jaccard = 0.93` over whole-file spans, because the framework-mandated declaration is most of each file (#331). The gate demotes them on measured content evidence while leaving a genuine reorder clone — real byte agreement — at act-now confidence. Both halves are pinned together: `crates/deslop/tests/lsh_only_nearmiss_recall.rs` asserts the genuine pair keeps `fused ≥ 0.85`, `crates/deslop/tests/issue_331_336_shape_only_saturation.rs` asserts the scaffold family does not.
+**Row 4 is routed on spread, not on content.** It reaches an act-now bucket on token overlap *alone*, and two shapes of cluster carry that estimate without earning the verdict. A **cross-file spread** (3+ members over 3+ files) is the #134 scaffolding pattern arriving through the token door instead of the structural one: six distinct Flutter widgets whose `build` bodies share nothing measure `structural = 0.00, token_jaccard = 0.93` over whole-file spans, because the framework-mandated declaration is most of each file (#331). An **unmeasured** cluster is one the content pass could not compare at all — the anchored routes may take one on trust because their Merkle equality is itself proof, but row 4 has no such signal, so unmeasured there means nothing is known (#108's `structural = 0.00, token_jaccard = 0.96` JSON-schema pair). Both demote to `LooselySimilar`, which the renderer hides — never `StructuralOnly`, which would claim a shape match `structural = 0.00` says does not exist.
+
+A measured *pair* is left alone even at low agreement, because that is the renamed Type-3 clone the content gate is structurally unable to vouch for ([fusion.md §FUSION-CONTENT-GATE](fusion.md#fusion-content-gate)). Both directions are pinned together: `crates/deslop/tests/lsh_only_nearmiss_recall.rs` asserts a genuine LSH-only pair keeps `fused ≥ 0.85` and `cli/detection.rs::detects_type3_clone_in_csharp_fixture` asserts the renamed C# pair still surfaces, while `crates/deslop/tests/issue_331_336_shape_only_saturation.rs` and `crates/deslop-core/tests/issue_98_99_108_120_122_thresholds.rs` assert the scaffold family and the unmeasured noise do not.
 
 Route 2 is why `token_jaccard < 0.05` is **not** a property of the bucket: the renamed-loop clones in `js-type2-loop` / `ts-type2-loop` are genuine cross-file Type-2 clones that land here with `token_jaccard = 1.00`. The invariant that does hold on both routes is that a `StructuralOnly` cluster carries no semantic support and stays below the act-now fused line; `common::signals::assert_structural_only_contract` is the shared assertion.
